refactor(file structure): move all api modules to base module
