+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/users/views.py
@@ -1,55 +1,30 @@
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.messages.views import SuccessMessageMixin
-from django.db.models import QuerySet
-from django.urls import reverse
-from django.utils.translation import gettext_lazy as _
-from django.views.generic import DetailView
-from django.views.generic import RedirectView
-from django.views.generic import UpdateView
+from rest_framework import status
+from rest_framework.decorators import action
+from rest_framework.mixins import ListModelMixin
+from rest_framework.mixins import RetrieveModelMixin
+from rest_framework.mixins import UpdateModelMixin
+from rest_framework.response import Response
+from rest_framework.viewsets import GenericViewSet
 
 from {{ cookiecutter.project_slug }}.users.models import User
 
+from .serializers import UserSerializer
 
-class UserDetailView(LoginRequiredMixin, DetailView):
-    model = User
+
+class UserViewSet(RetrieveModelMixin, ListModelMixin, UpdateModelMixin, GenericViewSet):
+    serializer_class = UserSerializer
+    queryset = User.objects.all()
     {%- if cookiecutter.username_type == "email" %}
-    slug_field = "id"
-    slug_url_kwarg = "id"
+    lookup_field = "pk"
     {%- else %}
-    slug_field = "username"
-    slug_url_kwarg = "username"
+    lookup_field = "username"
     {%- endif %}
 
+    def get_queryset(self, *args, **kwargs):
+        assert isinstance(self.request.user.id, int)
+        return self.queryset.filter(id=self.request.user.id)
 
-user_detail_view = UserDetailView.as_view()
-
-
-class UserUpdateView(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
-    model = User
-    fields = ["name"]
-    success_message = _("Information successfully updated")
-
-    def get_success_url(self) -> str:
-        assert self.request.user.is_authenticated  # type guard
-        return self.request.user.get_absolute_url()
-
-    def get_object(self, queryset: QuerySet | None=None) -> User:
-        assert self.request.user.is_authenticated  # type guard
-        return self.request.user
-
-
-user_update_view = UserUpdateView.as_view()
-
-
-class UserRedirectView(LoginRequiredMixin, RedirectView):
-    permanent = False
-
-    def get_redirect_url(self) -> str:
-        {%- if cookiecutter.username_type == "email" %}
-        return reverse("users:detail", kwargs={"pk": self.request.user.pk})
-        {%- else %}
-        return reverse("users:detail", kwargs={"username": self.request.user.username})
-        {%- endif %}
-
-
-user_redirect_view = UserRedirectView.as_view()
+    @action(detail=False)
+    def me(self, request):
+        serializer = UserSerializer(request.user, context={"request": request})
+        return Response(status=status.HTTP_200_OK, data=serializer.data)
